LPS-202141 move to permission util package
